Create new task now functional.
However, the response data is not successfully being pushed to the parent component
(task index)'s state. I was attempting to do this by passing an arrow function
(addNewTask) that sets taskIndex's state downstream and using that on a successful
post. But calling this.props.addNewTask() with the response data seems to throw an
error. Will get to the bottom of it. Just you wait.